Added my LCD driver code, driver works BUT not LCD
Reassigned pin assignments for the LCD to work with the PIC32MX360XXXL as per my table below.
LCD driver code is doing the the right things and setting the right pins to be 0's or 1's as and when needed.
But the LCD itself does not seem to be responding to these inputs. So there's a problems either with timing or something else, need to investigate.